requirements.txt
Django==4.2.17
django-oauth-toolkit==3.1.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
cryptography==44.0.0
